addition and subtraction implemented
multiplication stuff done with duals

arithmetic operations done

implemented unary negation

more dual num progress

backup

minor cleanups

backup

more fixes

backup

backup with more progress

more progress

most stuff with Dual done

demonstrating Scalar trait on Dual

initial prototype working

backu

adding some docs

new todos, some docs fixes

backup

more todos

more progress

almost there on the vector side, Vectorf64 should also be a trait

some more progress, still stuck

close to working vresion

basic version working after specializing to DVector

backup with slightly more progress

backup

backup for now

dual vec implemented

backup with some gradient progress

even more gradient progress

autodiff gradient working

gradient passing with two different types

gradient tests all passing

added a TODO

more TODOs, versioning updates

added some docs to dual_vector

more testing

backup

gradient actually working with multiple types

got vector valued deriv working with multiple types (hopefully)

more docs for get_sderivative

starting to fill out scalar-valued derivative tests

updated order of args

backup

still stuck

poly tests done

square root test done

tested product and quotient rules

hit another stopping point

got this thing working

exponential tests added

backup

power improvement

up to sin

backup for tonight

implementation of trig trait for Dual num

sin test

more trig tests

csc tests done

sec tested

finished testing trig functions

basic inverse trig tests

done with inverse trig tests

renamed test funcs

hyperbolic functions done

half of inverse hyperbolic functions done

finished inverse hyperbolic tests

renamed something

more progres

more docs updates

more progress

some more progress on docs

got rid of a TODO

updated to do conditional compilation on unit tests

added GetDualPart trait

backup

minor progress

added some placeholder tests

more unit test progress

unit tested NumCast

testing unary negation

backup

more progress

backup

more progress

is_normal checks done

tested classify

more fixes

testing floor

more dual number tests

trunc tested

fract tested

tested abs

more updates

more progress on dual

tested mul add

testing powi

more updates

implementing compare trait for dual just for testing

powf test domne

more powf tests

finished up powf tests

adding placeholder tests

got through remassign dual dual

another set of tests

implement more stuff for dual

lots of tests completed

improved a test

product rule test for central difference

improved product rule test

added quotient rule test for scalar valued derivative

finished new central diff tests

forward difference tests done

cleaned up autodiff product rule test

quotient rule cleaned up

another chain rule test cleaned up

finished chain rule cleanups

some TODO cleanups

tested sin

tested cos

first few trig functions tested

tested asin

standard trig test cases doe

sin_cos tested

tested atan2

removed extra TODO

removed a TODO

more docs

more docs

got rid of some TODOs that have already been addressed

fixed a failing unit test

removed unused variable

adding more TODOs

cleaned up gradient autodiff tests

resolved one TODO, added another

initial tests

array1 test working

dual vector tests done

got rid of yet another TODO

cleaned up lib.rs docs

backup with some renamings

more fixes, got rid of more TODOs

more updates

some more tests

some minor docs fixes

another success

filling out more unit tests

another test done

log tested

log10 test done

two more tests

sinh tests

tested cosh

tested taanh

unit tests up to atanh

backup

backup

backup for now

testing max

min test

backup with min/max

max and min tested

backup

tested hypot

updated rem dual tests

more tests

backup

more progress

getting close to done

adding some more explanations

backup

more docs

final TODO addressed

updated description

updating tolerance on a test

relax tol again

fixed some docs

+++ src/test_utils.rs
@@ -1,4 +1,5 @@
 #[cfg(test)]
+#[cfg(feature = "trig")]
 use trig::Trig;
 
 #[cfg(test)]
@@ -17,8 +18,8 @@ pub(crate) fn sqrt_deriv(x: f64) -> f64 {
 }
 
 #[cfg(test)]
-pub(crate) fn power_deriv(b: i32, x: f64) -> f64 {
-    (b as f64).powf(x) * (b as f64).ln()
+pub(crate) fn power_deriv(b: f64, x: f64) -> f64 {
+    b.powf(x) * b.ln()
 }
 
 #[cfg(test)]
@@ -47,21 +48,25 @@ pub(crate) fn cos_deriv(x: f64) -> f64 {
 }
 
 #[cfg(test)]
+#[cfg(feature = "trig")]
 pub(crate) fn tan_deriv(x: f64) -> f64 {
     x.sec().powi(2)
 }
 
 #[cfg(test)]
+#[cfg(feature = "trig")]
 pub(crate) fn csc_deriv(x: f64) -> f64 {
     -x.csc() * x.tan()
 }
 
 #[cfg(test)]
+#[cfg(feature = "trig")]
 pub(crate) fn sec_deriv(x: f64) -> f64 {
     x.sec() * x.tan()
 }
 
 #[cfg(test)]
+#[cfg(feature = "trig")]
 pub(crate) fn cot_deriv(x: f64) -> f64 {
     -x.csc().powi(2)
 }
@@ -107,21 +112,25 @@ pub(crate) fn cosh_deriv(x: f64) -> f64 {
 }
 
 #[cfg(test)]
+#[cfg(feature = "trig")]
 pub(crate) fn tanh_deriv(x: f64) -> f64 {
     x.sech().powi(2)
 }
 
 #[cfg(test)]
+#[cfg(feature = "trig")]
 pub(crate) fn csch_deriv(x: f64) -> f64 {
     -x.csch() * x.coth()
 }
 
 #[cfg(test)]
+#[cfg(feature = "trig")]
 pub(crate) fn sech_deriv(x: f64) -> f64 {
     -x.sech() * x.tanh()
 }
 
 #[cfg(test)]
+#[cfg(feature = "trig")]
 pub(crate) fn coth_deriv(x: f64) -> f64 {
     -x.csch().powi(2)
 }
